Moved CBox_v3 transmon to different file
Purpose is to have both the QWG transmon and CBox_v3 transmon in the
same file.
